added ralph's image modification code to a notebook and reorged files
