+++ Pipfile
@@ -4,7 +4,7 @@ url = "https://pypi.org/simple"
 verify_ssl = true
 
 [scripts]
-start = "python main.py"
+start = "jupyter notebook"
 
 [dev-packages]
 
@@ -15,6 +15,7 @@ pandas = "*"
 matplotlib = "*"
 jupyter = "*"
 scipy = "*"
+scikit-image = "*"
 
 [requires]
 python_version = "3.7"
